+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2021 - As admin or organizer, delete all meetings if agenda not published
 ### Fixed
 - MV-195 - Print user planning in badge scan screen
 - MV-195 - Badge participation type background color